refactor mocks
